changed the query param to id only, and added hook to fetch post by id. also moved home and chosen-post components into pages folder
