+++ src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import ApplicationLayout from './components/applicationLayout/ApplicationLayout';
-import Home from './components/home/Home';
-import ChosenPost from './components/chosenPost/ChosenPost';
+import Home from './components/pages/home/Home';
+import ChosenPost from './components/pages/chosenPost/ChosenPost';
 
 const queryClient = new QueryClient();
 
@@ -14,7 +14,7 @@ const App = () => {
         <ApplicationLayout>
             <Routes>
               <Route path="/" element={<Home />} />
-              <Route path="/chosen-post/:encodedPost" element={<ChosenPost />} />
+              <Route path="/posts/:id" element={<ChosenPost />} />
             </Routes>
         </ApplicationLayout>
       </BrowserRouter>
